hw4/requirements.txt
numpy==2.2.0
Pillow==11.0.0
tqdm==4.67.1
